remove newly defined redundant constants with Network methods equivalents (#681)
